tools/mki-copy-pkgs: split processing of pkglists
mki_list_pkgs() is now used per package list (or package name),
and not in one fell swoop.

The rationale for this is that the more complex/versatile products can
include conflicting packages in different package lists (e.g. a few
MTAs) but if all pkglists are processed at once, we get conflicts in
that case.

I've stumbled upon that with ALTSP after base/disk subprofiles were
merged as main subprofile in m-p-d -- and had to introduce yet another
subprofile, ltsp, just to toss in a few more packages (some of them
conflicting with serverside) to build client chroot.

There _is_ a minor performance hit: measured on m-p's minicd.iso,
it's 5:41 instead of 5:39 on average (~0.5% of the whole time).
I tend to think it was an overoptimization though.