Revert "mki-pack-isoboot: added mkisofs -sort support"
This reverts commit 54c2f9096b7f31f7318cd4ef9b67eac70a4f2b21.

The reason is that xorriso doesn't support sortfiles
in a meaningful way and the main reason to introduce
that feature is done there in another way out-of-box.

As the ISO9660 producers are (hopefully) going to merge
and there was no visible interest in the feature which
only got used within mkimage-profiles, seems it's not
worth maintaining a kind of backwards compatibility.