mki-copy-efiboot: initial implementation
The purpose of this script is to prepare a FAT32 image
that holds the initial EFI/UEFI-compatible bootloader;
hybridization will turn this file into a "partition".

ELILO is one of the simplest ways to get the stone rolling
but it has an obvious deficiency when combined with UEFI spec
that promises only FAT as the definitely available filesystem:
one needs to provide a copy of kernel and initramfs there too
(unless one is going to fiddle with ISO9660 UEFI driver).