mki-copy-efiboot: copy EFI shell if requested
UEFI shell is a pretty valuable debugging and rescue tool.

Please note that while refind is happy to have it under
EFI/tools/, firmware implementations at hand would only
offer to run the shell sitting in the ESP root directory.

The trick is to put shellx64.efi under EFI/ so that it's
included in FAT image size calculations but then move it
aside that directory so that it ends up in FAT root.

The bonus trick is that firmware probably won't boot /our/
efi-shell-signed binary since MOK isn't available there,
so only shim+refind combo is reasonably expected to run it.