added preinstall subpackage
Linux 3.6 includes the code to limit user's ability to hardlink
non-owned files and it's now on by default on ALT Linux systems.

It doesn't make sense to try and work around the security measure
that's generally fine (we used to ship it back in 2.2/2.4 kernel
days by applying OW patch) as any attempt to do so would either
employ a privileged helper (might belong to hasher-priv) or imply
effectively destroying the permissions by resetting them to those
of the caller user (the one who runs mkimage and hasher).

There is a workaround of using full copies instead of hardlinks
but of course the resource usage increases dramatically (think
RAM for tmpfs builddir and time+CPU for disk-based one).

So let it be this way until a better one is devised, if ever:

  echo 0 > /proc/sys/fs/protected_hardlinks
  echo "fs.protected_hardlinks = 0" > /etc/sysctl.d/mkimage.conf

This subpackage is to provide the official way of setting up this
host environment for mkimage to function properly.

Yes it did take half a year to figure out, sorry for that.