spec: fix missing quoting
I managed to forget that shell evaluates *all* of the command
before running it; thus control reaches the wrong part when
there is no /proc/sys/fs/protected_hardlinks to cat(1).

Let's make it rugged and straightforward, and let's avoid the
needless action and message in the main package's %post either.

Use stderr for messages while at that.