mki-copy-efiboot: hardlink image contents
It has dawned on me that ISO9660 actually does hardlinks
(thanks boyarsh@: reading xorriso docs back then wasn't enough),
so let's reduce the number of copies from 3 to 2 (another one in
El Torito FAT image is not that easy to get rid of, need a driver
for ISO9660 and proper addressing since UEFI spec is brilliant
enough to mandate FAT32 but not ISO9660).

Hacking isolinux.cfg to use /EFI/BOOT paths works too
but that's not nearly as elegant.

-30M per image, on average.