mki-copy-efiboot: rework refind installation
The overall goal of what became this commit was to reduce
refind's image footprint given that we put two copies in
(one as a primary bootloader with limited number of boot
options and another one as a scanning multiboot tool).

While the subset of icons was made to save some El Torito
image space (which presented funny behaviour when going
over 32Mb, see also #29662), it was left that way upon
more consideration as most os_* icons aren't ever going
to be used in that limited configuration anyways.

Some more use of hardlinks was added -- this happened
before the overall postprocessing and I chose to leave
the implementation in place.  If someone decides that
it adds peculiarity to what should have been generic,
feel free to improve.