Added a new type of bootable ISO with grub-pc
Feature: Protective MBR + GPT is used. This should make life easier
for users. Now, after writing to a USB flash drive and correcting
the GPT table, users can freely dispose of the remaining free space.
This surprised me a lot, but this iso is successfully loaded on old
i686 PC with BIOS.

Signed-off-by: Anton Midyukov <antohami@altlinux.org>