mki-image-prepare: Allow to run script just after unpack IMAGE_INIT_LIST
In some cases, it is necessary to perform steps before installing packages
into the chroot, but before executing any scripts from rpm packages.
For example, you may want to copy /etc/passwd, /etc/group, /etc/login.defs
into the chroot to control the uid/gid inside the image.

To do this, the --predb-prog=SCRIPT option appeared in hsh-initroot. We
allow this option to be specified to initialize a working chroot.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>