tools/mki-pack-boot: rework boot modes support
In the beginning, the specification BOOT_TYPE=isolinux signified the
employment of isolinux as the ISO image's bootloader.  However, at a
later point, this approach was altered.  Currently, the setting
BOOT_TYPE=isolinux implies that the ISO image may or may not include
compatibility with EFI.  The introduction of BOOT_TYPE=grubpcboot
followed a similar approach: it utilizes grub-pc and similarly offers
uncertain support for EFI.  This interface is not well-suited for the
mkimage project and may be challenging to comprehend.

This modification introduces new boot options, namely 'isolinux-boot'
and 'grub-pc-boot', which do not imply EFI.  Following this change, it
is expected that a profile explicitly include 'efiboot' in the list of
required boot types.  For instance, one should specify
'BOOT_TYPE=isolinux-boot efiboot' and 'BOOT_TYPE=grub-pc-boot efiboot'
instead of 'BOOT_TYPE=isolinux' and 'BOOT_TYPE=grubpcboot' respectively.

The boot types 'isolinux' and 'grubpcboot' are supported for backward
compatibility, but both of them explicitly enable EFI support now.

Fixes: 0.2.5-alt1~2 ("mki-pack-{iso,}boot: added initial efiboot support")
Fixes: 0.2.39-alt1~2 ("Added a new type of bootable ISO with grub-pc")

Signed-off-by: Gleb Fotengauer-Malinovskiy <glebfm@altlinux.org>