More accurately copy environ variables
The printenv utility prints the names and values of env variables.
Variable values are not escaped in any way. Therefore, we cannot use
the output of this utility directly.

Problematic example:

 $ export INFO_foo="1
 > INFO_bar=2
 > INFO_baz=3"

 $ printenv | sed -ne 's/^\(\(GLOBAL\|INFO\)_[^=]\+\)=/\1\t/p'
 INFO_foo        1
 INFO_bar        2
 INFO_baz        3

To avoid this problem, we only need to get the names of the environment
variables. Unfortunately, neither printenv nor env are able to show only
names. Therefore, we break the values of variables (we do not need them)
and remove newlines. After that, we can return newlines to the end of
each environment variable. Now we can properly truncate the name in the
each string.

Having the names of the variables, we can safely get their values or
remove them from the environment.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>