Add mki-chroot-sh-functions
This shell functions file can be included by any script that is executed
within the chroot environment.  This approach seems to be the simplest
and most straightforward solution for consolidating common code segments
found in the chrooted scripts, which are implemented using heredoc
redirections.  The implementation of the actual common code segments is
planned for future commits.

Signed-off-by: Gleb Fotengauer-Malinovskiy <glebfm@altlinux.org>