mki-pack-*: add option '-joliet-long' for mkisofs
joliet is more widely supported than Rock Ridge. Microsoft Windows
or the 7z program read joliet, but not Rock Ridge. Some RPM packages
on disk are name longer than 64 characters, which causes names to be
truncated in joliet. By increasing the length of names to 103 characters,
we will solve the problem with unpacking iso in MS Windows and the
7z program.

Signed-off-by: Anton Midyukov <antohami@altlinux.org>