mki-copy-grub-efi-chrooted: Capitalize primary bootloader filename
... for non-secureboot platforms.

Section 3.5.1.1 of the UEFI specification[1] mandates usage of
capital latters by the EFI firmware for generating the default
file name for removable media. Failing to adhere this requirement
renders our images unbootable on *some* loongarch64 and riscv64
machines.

This commit does not touch the Intel platforms, we're not aware
of any actual issues there and interactions with secure boot
require further investigations and testing.

[1] https://uefi.org/specs/UEFI/2.10_A/03_Boot_Manager.html#removable-media-boot-behavior

Signed-off-by: Ivan A. Melnikov <iv@altlinux.org>
Reported-by: Dmitriy Terekhin <jqt4@altlinux.org>